Refactor QRCodeGenerator into separate files
