+++ QRCoder/QRCodeGenerator.AlignmentPattern.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace QRCoder
+{
+    public partial class QRCodeGenerator
+    {
+        private struct AlignmentPattern
+        {
+            public int Version;
+            public List<Point> PatternPositions;
+        }
+    }
+}
